During the transition, a number of files names have been changed to a different case. These have been added to the vcproj file.
This poses a problem with subversion. latexproject.cpp will be renamed by Studio and Subversion does not recognize it anymore. While I like the new names, we cannot afford these problems. So unfortunately I had to re-add these files to the vcproj with original case. Sorry.

If we figure out a nice way around this problem, lets go for better filenames with consistent case.