MakeIndex uses the %tm instead of the %bm placeholder for output files
Due to stronger security mechanisms, the BibTeX and MakeIndex versions
distributed starting with MiKTeX 2.9 and TeX Live 2010 do not allow absolute
paths for output files anymore.

This changeset also adds the -t option to default MakeIndex arguments. If the
option is not specified, MakeIndex complains that it can't create the transcript
file.